refactor: Removed interface suffix from abstraction files.
