+++ src/app/core/abstractions/entity.parser.ts
