+++ src/app/core/core.module.ts
@@ -9,10 +9,10 @@ import { AdLibPieceService } from './services/ad-lib-piece.service'
 import { ConnectionStatusObserver } from './services/connection-status-observer.service'
 import { RundownEventObserver } from './services/rundown-event-observer.service'
 import { RundownStateService } from './services/rundown-state.service'
-import { RundownEventParser } from './abstractions/rundown-event-parser.interface'
+import { RundownEventParser } from './abstractions/rundown-event.parser'
 import { BasicRundownStateService } from './services/basic-rundown-state.service'
-import { BasicRundownService } from './abstractions/basic-rundown-service'
-import { EntityParser } from './abstractions/entity-parser.interface'
+import { BasicRundownService } from './abstractions/basic-rundown.service'
+import { EntityParser } from './abstractions/entity.parser'
 import { ZodEntityParser } from './parsers/zod-entity-parser.service'
 import { ZodRundownEventParser } from './parsers/zod-rundown-event-parser.service'
