+++ src/app/core/services/http-basic-rundown.service.ts
@@ -2,9 +2,9 @@ import { Injectable } from '@angular/core';
 import { HttpClient} from '@angular/common/http';
 import { HttpErrorService } from './http-error.service';
 import { catchError, map, Observable } from 'rxjs'
-import { BasicRundownService } from '../abstractions/basic-rundown-service';
+import { BasicRundownService } from '../abstractions/basic-rundown.service';
 import { BasicRundown } from "../models/basic-rundown";
-import { EntityParser } from '../abstractions/entity-parser.interface'
+import { EntityParser } from '../abstractions/entity.parser'
 
 // TODO: Change when we get a setup for handling configurations.
 const RUNDOWN_URL: string = 'http://localhost:3005/api/rundowns'
