+++ src/app/core/services/http-rundown.service.spec.ts
@@ -2,7 +2,7 @@ import { HttpRundownService } from './http-rundown.service';
 import { HttpErrorService } from './http-error.service'
 import { instance, mock } from '@typestrong/ts-mockito'
 import { HttpClient } from '@angular/common/http'
-import { EntityParser } from '../abstractions/entity-parser.interface'
+import { EntityParser } from '../abstractions/entity.parser'
 
 describe('HttpRundownService', () => {
   it('should be created', () => {
