+++ src/app/event-system/abstractions/event-observer.service.ts
